fix: auth check delete page

+++ apps/web/src/pages/api/page/unpublish.ts
@@ -1,8 +1,10 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { supabase } from "../../../lib/supabase/supabase";
+import { DIDSession } from "did-session";
 
 type Payload = {
   pageId: string;
+  ceramicSession: string;
 };
 
 export default async function handler(
@@ -10,12 +12,21 @@ export default async function handler(
   res: NextApiResponse
 ) {
   try {
-    const { pageId } = JSON.parse(req.body) as Payload;
+    const { pageId, ceramicSession } = JSON.parse(req.body) as Payload;
+
+    const session = await DIDSession.fromSession(ceramicSession);
+
+    if (!session.hasSession || session.isExpired) {
+      return res.status(401).json({ success: false });
+    }
+
+    const publisherAddress = session.id.split(":")[4];
 
     const result = await supabase
       .from("page_publication")
       .delete()
-      .eq("page_id", pageId);
+      .eq("page_id", pageId)
+      .eq("publisher_address", publisherAddress);
 
     if (result.error) {
       console.error(result.error);
